Uncouple semaphore implementation from requests
- create package semimpl
- remove unnecessary request.DoUntil
