+++ cmd/runner/main.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"time"
 
-	"github.com/benchttp/runner/record"
 	"github.com/benchttp/runner/request"
 )
 
@@ -51,12 +50,7 @@ func main() {
 	ctx, cancel := context.WithTimeout(context.Background(), duration)
 	defer cancel()
 
-	var rec []record.Record
-	if requests > 0 {
-		rec = request.Do(ctx, requests, concurrency, url, timeout)
-	} else {
-		rec = request.DoUntil(ctx, concurrency, url, timeout)
-	}
+	rec := request.Do(ctx, requests, concurrency, url, timeout)
 
 	fmt.Println("total:", len(rec))
 }
